ハンドル管理・アロケータのリファクタリングと型安全化
ハンドルとアロケータの実装を大幅に刷新し、型安全なHandle/Allocatorテンプレートを導入しました。これにより、リソース管理の効率性・安全性・拡張性が向上しています。VulkanDeviceは新しいハンドルアロケータを利用するよう変更され、Swapchainのセマフォ数型も修正されています。

+++ engine/src/rhi/vulkan/VulkanSwapchain.cpp
@@ -174,7 +174,7 @@ void VulkanSwapchain::createFrameContext()
         vkCreateSemaphore(device, &semaphoreCI, nullptr, &frame.renderComplete);
     }
 
-    uint32_t presentCompleteSemaphoreCount = m_images.size() + 1;
+    size_t presentCompleteSemaphoreCount = m_images.size() + 1;
     m_presentSemaphoreList.resize(presentCompleteSemaphoreCount);
     for (auto& semaphore : m_presentSemaphoreList) {
         VkSemaphoreCreateInfo semaphoreCI{
